Remove "OscilSmp/" and "FMSmp/" from ADvoice (finally!)
This removes the ports "OscilSmp/" and "FMSmp/" from ADnoteVoiceParam.
They are refering to non-RT objects only.

While the ports are removed, the objects OscilGn and FmGn remain in
ADnoteVoiceParam. This is probably just to make XML load/save easier,
or to keep consistency (MW also did not store PAD note's non RT params,
it only kept pointers).

A workaround may still be needed for automation: If a host sends non-RT
messages to master, they must somehow be forwarded to MiddleWare.