Let MW generate variable sized Tensor3s
Previously to this commit, AD note had started with a WaveTable with the
maximum possible number of random seeds. This was the capacity, and the
size could change without allocations.

This introduced memory issues, e.g. with the PluginTests allocating more
than 16 GB of RAM on common machines.

This commit introduces changes:

* There is no difference between capacity and size in the WaveTable
  anymore (don't look at the horrible changes in WaveTable.h)
* AD note initially starts with only one random seed (because the
  default wave is a random-less sine)
* MW allocates a `Tensor3` of fitting size on each parameter change and
  sends this `Tensor3` to ADnoteParameters